fixed diffuse

+++ src/Cell.java
@@ -9,8 +9,8 @@ public class Cell {
 
 	private int x;
 	private int y;
-	private int maxGrain;
-	private int grain;
+	private double maxGrain;
+	private double grain;
 	private int currentInterval = 0;
 	private int tickInterval = Configuration.gainGrowthInterval;
 	private int grainGrowth = Configuration.numGrainGrown;
@@ -34,19 +34,19 @@ public class Cell {
 		return new Vector2<Integer>(getX(), getY());
 	}
 
-	public int getGrain() {
+	public double getGrain() {
 		return this.grain;
 	}
-	public void setInitialGrain(int value) {
+	public void setInitialGrain(double value) {
 		this.grain = value;
 	}
-	public void setMaxGrain(int value) {
+	public void setMaxGrain(double value) {
 		this.maxGrain = value;
 	}
-	public int getMaxGrain() {
+	public double getMaxGrain() {
 		return this.maxGrain;
 	}
-	public void harvestGrain(int value) {
+	public void harvestGrain(double value) {
 		this.grain -= value;
 	}
